Create random_number.py
This program is educational - guessing numbers. the program generates a random number in the range from 11 to 100 and asks the user to guess this number. If the user's guess is greater than a random number, then the program should output the message text_1. If the guess is less than a random number, then the program should output the message text_2.  If the user guesses the number, the program should congratulate him and output the message text_3.